successfully retrieved steam scores
